# dbmongo/docs/swagger/swagger.yaml
components:
  schemas:
...
    Purge:
      properties:
        RawData:
          properties:
            info:
              type: string
            error:
              type: string
        Journal:
          properties:
            info:
              type: string
            error:
              type: string
        Features:
          properties:
            info:
              type: string
            error:
              type: string
        Public:
          properties:
            info:
              type: string
            error:
              type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
info:
  contact: 
    name: "contact support"
    email: "contact@signaux-faibles.beta.gouv.fr"
  description: |-
    Cette API centralise toutes les fonctionnalités du module de traitement de données [openSignauxFaibles](https://github.com/entrepreneur-interet-general/opensignauxfaibles)  
    Ce [projet](https://beta.gouv.fr/startups/signaux-faibles.html) fournit une plateforme technique à l'exploitation des signaux faibles pour anticiper les difficultés des entreprises  
  
    La plateforme fournit les services suivants    
    - Authentification  
    - Gestion des batches d'intégration  
    - Lancement/ordonnancement des traitements  
    - Consultation des données  
    
    Toutes les erreurs rencontrées lors du traitement seront signalées par le statut `HTTP 500` assorti du message d'erreur ayant provoqué le défaut du traitement.  
    Par ailleurs, tous les services authentifiés sont succeptibles de renvoyer le statut `HTTP 401` dans le cas où l'authentification est erronée.  Le code `HTTP 403` est également communiqué lorsque les attributions sont insuffisantes pour traiter la requête.  
    Dans le cas où une requête est malformée, le code `HTTP 400` sera renvoyé assorti d'un message indiquant l'erreur rencontrée dans l'analyse du corps de la requête  
  license:
    name: Licence MIT
    url: https://raw.githubusercontent.com/entrepreneur-interet-general/opensignauxfaibles/master/LICENSE
  title: API openSignauxFaibles
  version: "1.1"
paths:
...
  /api/data/compact:
    get:
      summary: compacter la base de données
      description: |-
        Ce traitement permet le compactage de la base de données.  
        Ce compactage a pour effet de réduire tous les objets en clé uniques comportant dans la même arborescence toutes les données en rapport avec ces clés.  
        Ce traitement est nécessaire pour rendre disponible les données à /api/data/reduce les nouvelles données importées.
      responses:
        "200":
          description: Le traitement s'est bien déroulé
          content: 
            application/json:
              schema:
                type: string
              example: "ok"
      security:
      - BearerAuth: []
      tags:
      - Traitements
...
  /api/data/purge:
    get:
      description: |-
        Suppression de tous les objets de données brutes contenus dans les collections RawData, Features, Journal et Public  
        **Ceci est une opération de remise à zéro de la base de données**  
        **Seule la configuration des batches est conservée**  
        **Cette opération est irréversible et dangereuse et ne devrait être utilisée qu'à des fins de développement**  
        
        **!!! SI VOUS NE SAVEZ PAS CE QUE VOUS FAITES ICI, PASSEZ VOTRE CHEMIN !!!**
      responses:
        "200":
          description: statistiques de suppression
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purge"
      security:
      - BearerAuth: []
      summary: remettre à zéro les données /!\
      tags:
      - Traitements
...
openapi: "3.0.0"